// lab4-v3-demo-api/student_frontend/types/student.ts
export interface Student {
    id: string;
    firstName: string;
    lastName: string;
    studentNumber: string;
    email: string;
    course: string;
  }
  
  export interface StudentCreate {
    firstName: string;
    lastName: string;
    studentNumber: string;
    email: string;
    course: string;
  }
  